Reorganize menu related files
