+++ client/src/data/prices.json
@@ -1,0 +1,5 @@
+{
+    "Samosa": 2.99,
+    "Pakora": 3.79,
+    "Gulab Jamun": 4.99
+}
