+++ client/src/menu/MenuItem.js
@@ -1,9 +1,9 @@
-function MenuItem({ name, image, description }) {
+function MenuItem({ name, imagePath, imageAlt, description }) {
   return (
     <div className="menu-item">
       <h3 className="menu-item-name">{name}</h3>
       <div className="menu-img-container">
-        <img className="menu-img" src={image.path} alt={image.alt} />
+        <img className="menu-img" src={imagePath} alt={imageAlt} />
       </div>
       <p>{description}</p>
     </div>
